refactor: move the valhalla use in handler
