+++ asgard/CMakeLists.txt
@@ -24,7 +24,10 @@ add_custom_command(OUTPUT ${PROTO_HDRS} ${PROTO_SRCS}
 
 set_source_files_properties(${PROTO_HDRS} ${PROTO_SRCS} PROPERTIES GENERATED TRUE)
 
-
-
-add_executable(asgard asgard.cpp ${CMAKE_SOURCE_DIR}/utils/zmq.cpp ${PROTO_SRCS})
+add_executable(asgard
+  asgard.cpp
+  handler.cpp
+  ${CMAKE_SOURCE_DIR}/utils/zmq.cpp
+  ${CMAKE_SOURCE_DIR}/utils/coord_parser.cpp
+  ${PROTO_SRCS})
 target_link_libraries(asgard boost_system boost_regex boost_thread ${BOOST_DEV_LIBS} valhalla z curl zmq protobuf) #TODO do not hardcode lib name
